+++ R/plot_annual_outside_normal.R
@@ -110,10 +110,11 @@ plot_annual_outside_normal <- function(data = NULL,
   normal_plots <- dplyr::mutate(normal_plots,
                                 plot = purrr::map2(data, STATION_NUMBER, 
         ~ggplot2::ggplot(data = ., ggplot2::aes(x = Year, y = Value, color = Statistic)) +
-          ggplot2::geom_line(alpha = 0.5) +
-          ggplot2::geom_point() +
+          ggplot2::geom_line(alpha = 0.5, na.rm = TRUE) +
+          ggplot2::geom_point(na.rm = TRUE) +
           ggplot2::facet_wrap(~Statistic, scales="free_x", ncol = 1, strip.position = "right") +
-          ggplot2::scale_x_continuous(breaks = scales::pretty_breaks(n = 6)) +
+          ggplot2::scale_x_continuous(breaks = scales::pretty_breaks(n = 8))+
+          {if(length(unique(normal_data$Year)) < 8) ggplot2::scale_x_continuous(breaks = unique(normal_data$Year))}+
           ggplot2::scale_y_continuous(breaks = scales::pretty_breaks(n = 6)) +
           ggplot2::ylab("Number of Days") +
           ggplot2::xlab("Year") +