+++ R/plot_monthly_cumulative_stats.R
@@ -167,11 +167,11 @@ plot_monthly_cumulative_stats <- function(data = NULL,
                                                "25th-75th Percentile" = "skyblue1", "75th-95th Percentile" = "dodgerblue2",
                                                "95th Percentile-Max" = "royalblue4")) +
          ggplot2::scale_color_manual(values = c("Median" = "purple3", "Mean" = "springgreen4")) +
-         {if(!log_discharge) ggplot2::scale_y_continuous(expand = c(0, 0))} +
-         {if(log_discharge) ggplot2::scale_y_log10(expand = c(0, 0))} +
-         {if(log_discharge) ggplot2::annotation_logticks(base= 10, "left", colour = "grey25", size = 0.3,
-                                                         short = ggplot2::unit(.07, "cm"), mid = ggplot2::unit(.15, "cm"),
-                                                         long = ggplot2::unit(.2, "cm"))} +
+         {if (!log_discharge) ggplot2::scale_y_continuous(expand = c(0, 0), breaks = scales::pretty_breaks(n = 7))} +
+         {if (log_discharge) ggplot2::scale_y_log10(expand = c(0, 0), breaks = scales::log_breaks(n = 8, base = 10) )} +
+         {if (log_discharge) ggplot2::annotation_logticks(base= 10, sides = "l", colour = "grey25", size = 0.3,
+                                                          short = ggplot2::unit(.07, "cm"), mid = ggplot2::unit(.15, "cm"),
+                                                          long = ggplot2::unit(.2, "cm"))} +
          ggplot2::xlab("Month")+
          ggplot2::scale_x_discrete(expand = c(0.01,0.01)) +
          {if(!use_yield) ggplot2::ylab("Cumulative Discharge (cubic metres)")} +